requirements.txt
Flask==2.3.0
Flask-MySQLdb==1.0.1
Werkzeug==2.3.0
pymysql==1.1.0
